refactor: separate project container and lnb
