+++ apps/web/src/services/project/ProjectContainer.vue
@@ -1,0 +1,32 @@
+<script setup lang="ts">
+import CenteredPageLayout from '@/common/modules/page-layouts/CenteredPageLayout.vue';
+import GeneralPageLayout from '@/common/modules/page-layouts/GeneralPageLayout.vue';
+import VerticalPageLayout from '@/common/modules/page-layouts/VerticalPageLayout.vue';
+
+import ProjectLNB from '@/services/project/ProjectLNB.vue';
+</script>
+
+<template>
+    <fragment>
+        <vertical-page-layout v-if="$route.meta.lnbVisible"
+                              :breadcrumbs="undefined"
+        >
+            <template #sidebar>
+                <project-l-n-b />
+            </template>
+            <template #default>
+                <router-view />
+            </template>
+        </vertical-page-layout>
+        <centered-page-layout v-else-if="$route.meta.centeredLayout"
+                              has-nav-bar
+        >
+            <router-view />
+        </centered-page-layout>
+        <general-page-layout v-else
+                             :breadcrumbs="undefined"
+        >
+            <router-view />
+        </general-page-layout>
+    </fragment>
+</template>
